Product

+++ AdminSite/DataAccess/Product/IProduct.cs
@@ -5,10 +5,10 @@ namespace AdminSite.DataAccess.Product
     public interface IProduct
     {
         public int ProductId { get; }
-        public byte Category { get; }
+        public int Category { get; }
         public int Quantity { get; }
         public string Name { get; }
-        public int Price { get; }
+        public decimal Price { get; }
         public Image Img { get; }
         public string Description { get; }
